WIP conversion to Active storage, basic function is working
